+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed some UI layout issues in the Conversation screen composer and in the Scheduling confirmation screen.
 
 ## [1.0.0-alpha28] - 2022-12-13